refactor(Column, Height): 패키지 구조 변경
