+++ src/main/java/laddergame/domain/ladder/Ladder.java
@@ -1,5 +1,7 @@
 package laddergame.domain.ladder;
 
+import laddergame.domain.vo.Height;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
